requirements.txt
alembic==1.0.3
argh==0.26.2
astroid==2.1.0
autopep8==1.4.3
Click==7.0
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Admin==1.5.2
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.11
pathtools==0.1.2
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.0
python-dateutil==2.7.5
python-editor==1.0.3
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.14
tablib==0.12.1
unicodecsv==0.14.1
watchdog==0.9.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
xlwt==1.3.0
